# optimize traffic road class slicing