### Description
<!--
Include issue references (e.g., fixes [#issue](link))
Include necessary implementation details (e.g. I opted to use this algorithm because ... and test it in this way ...).
-->
Refs https://github.com/mapbox/mapbox-navigation-android/issues/4096#issuecomment-816771054.

After applying the changes from the PR, I see below values for a route from Wrocław in Poland to Seville in Spain:

|   | generating a route | selecting an alternative |
| ------------- | ------------- | ------------- |
| Old impl  |  ~10.5s | ~24.5s  |
| New impl  | ~3.5s | ~6s  |

Those measurements are before applying the changes from https://github.com/mapbox/mapbox-navigation-android/pull/4209, which results in:
|   | generating a route | selecting an alternative |
| ------------- | ------------- | ------------- |
| New impl + #4209  | ~0.3s  | ~0.2s  |

This contribution doesn't change any business logic, it just refactors the method of finding out the road class name. I think there's something off with the index we're using, but let's investigate that separately.

On top of the above measurements, we have below 0.5s of rendering latency.

### Changelog
<!--
Include changelog entry (e.g. Fixed an unexpected change in recenter button when resuming the app.).
See https://github.com/mapbox/navigation-sdks/blob/main/documentation/android-changelog-guidelines.md.
You can remove the changelog block and add a `skip changelog` label, when applicable.
 -->
```
<changelog>Improved performance of `MapboxRouteLine`.</changelog>
```

<!--
---------- CHECKLIST ----------
1. Add related labels (`bug`, `feature`, `new API(s)`, `SEMVER-MAJOR`, `needs-backporting`, etc.).
2. Update progress status on the project board.
3. Request a review from the team, if not a draft.
4. Add targeted milestone, when applicable.
5. Create ticket tracking addition of public documentation pages entry, when applicable.
-->