@LukasPaczos 2021-04-09
Looks like unit tests have me in check for that statement about the index! Fixing it up right now.