@Guardiola31337 2021-04-09
`Routeline and Route Arrow Example` is crashing for me 👀 

```
2021-04-09 12:56:50.619 25383-25383/com.mapbox.navigation.examples E/AndroidRuntime: FATAL EXCEPTION: main
    Process: com.mapbox.navigation.examples, PID: 25383
    java.lang.IndexOutOfBoundsException: Empty list doesn't contain element at index 0.
        at kotlin.collections.EmptyList.get(Collections.kt:36)
        at kotlin.collections.EmptyList.get(Collections.kt:24)
        at com.mapbox.navigation.examples.core.MapboxRouteLineAndArrowActivity$initListeners$1.onClick(MapboxRouteLineAndArrowActivity.kt:468)
        at android.view.View.performClick(View.java:7140)
        at android.view.View.performClickInternal(View.java:7117)
        at android.view.View.access$3500(View.java:801)
        at android.view.View$PerformClick.run(View.java:27351)
        at android.os.Handler.handleCallback(Handler.java:883)
        at android.os.Handler.dispatchMessage(Handler.java:100)
        at android.os.Looper.loop(Looper.java:214)
        at android.app.ActivityThread.main(ActivityThread.java:7356)
        at java.lang.reflect.Method.invoke(Native Method)
        at com.android.internal.os.RuntimeInit$MethodAndArgsCaller.run(RuntimeInit.java:492)
        at com.android.internal.os.ZygoteInit.main(ZygoteInit.java:930)
2021-04-09 12:56:50.621 25383-25486/com.mapbox.navigation.examples E/FirebaseCrashlytics: Tried to write a fatal exception while no session was open.
```

https://user-images.githubusercontent.com/1668582/114215469-a2e17b00-9933-11eb-85b1-8f28d2b903ea.mp4