@LukasPaczos 2021-04-09
@Guardiola31337 looks like the crash is from the example app - you clicked start navigation before the route was loaded. I pushed some hardening anyway though.