@Guardiola31337 2021-04-09
> looks like the crash is from the example app - you clicked start navigation before the route was loaded. I pushed some hardening anyway though.

Yeap, thought about that but wait for some time and the route line didn't appear at all 🤔