@zugaldia 2021-04-09
> route from Wrocław in Poland to Seville in Spain:

This is all good but next time you need use Granada, Spain as your destination or I'll block your PR.