@cafesilencio 2021-04-09
That activity failure is occurring because of a previous merge mistake.  The call to MapboxNavigation.setRoutes got removed.  I can add it back in a separate PR.